requirements.txt
pytorch
numpy
pandas
ipdb
tqdm
nltk
stanfordcorenlp
sklearn
